feat: renamed folder from home -> index
