+++ src/pages/index/Terminal.jsx
